DDCNL-1725 | AddressSelectorController created
