+++ app/controllers/AddressController.scala
@@ -136,93 +136,6 @@ class AddressController @Inject()(
       }
     }
 
-  def showAddressSelectorForm(typ: AddrType) =
-    authenticate.async { implicit request =>
-      gettingCachedJourneyData(typ) { journeyData =>
-        journeyData.recordSet match {
-          case Some(set) =>
-            Future.successful(
-              Ok(
-                views.html.personaldetails.addressSelector(
-                  AddressSelectorDto.form,
-                  set,
-                  typ,
-                  postcodeFromRequest,
-                  filterFromRequest
-                )
-              )
-            )
-          case _ => Future.successful(Redirect(address.routes.PostcodeLookupController.onPageLoad(typ)))
-        }
-      }
-    }
-
-  def processAddressSelectorForm(typ: AddrType): Action[AnyContent] =
-    authenticate.async { implicit request =>
-      val errorPage = Future.successful(
-        InternalServerError(
-          views.html.error(
-            "global.error.InternalServerError500.title",
-            Some("global.error.InternalServerError500.title"),
-            List("global.error.InternalServerError500.message")
-          )
-        )
-      )
-
-      addressJourneyEnforcer { _ => personDetails =>
-        gettingCachedJourneyData(typ) { journeyData =>
-          AddressSelectorDto.form.bindFromRequest.fold(
-            formWithErrors => {
-
-              journeyData.recordSet match {
-                case Some(set) =>
-                  Future.successful(
-                    BadRequest(
-                      views.html.personaldetails.addressSelector(
-                        formWithErrors,
-                        set,
-                        typ,
-                        postcodeFromRequest,
-                        filterFromRequest
-                      )
-                    ))
-                case _ =>
-                  Logger.warn("Failed to retrieve Address Record Set from cache")
-                  errorPage
-              }
-            },
-            addressSelectorDto => {
-              journeyData.recordSet
-                .flatMap(_.addresses.find(_.id == addressSelectorDto.addressId.getOrElse(""))) match {
-                case Some(addressRecord) =>
-                  val addressDto = AddressDto.fromAddressRecord(addressRecord)
-
-                  for {
-                    _ <- addToCache(SelectedAddressRecordId(typ), addressRecord)
-                    _ <- addToCache(SubmittedAddressDtoId(typ), addressDto)
-                  } yield {
-                    val postCodeHasChanged = !postcodeFromRequest
-                      .replace(" ", "")
-                      .equalsIgnoreCase(personDetails.address.flatMap(_.postcode).getOrElse("").replace(" ", ""))
-                    (typ, postCodeHasChanged) match {
-                      case (PostalAddrType, false) =>
-                        Redirect(routes.AddressController.showUpdateAddressForm(typ))
-                      case (_, true) => Redirect(routes.AddressController.enterStartDate(typ))
-                      case (_, false) =>
-                        addToCache(SubmittedStartDateId(typ), DateDto(LocalDate.now()))
-                        Redirect(routes.AddressController.reviewChanges(typ))
-                    }
-                  }
-                case _ =>
-                  Logger.warn("Address selector was unable to find address using the id returned by a previous request")
-                  errorPage
-              }
-            }
-          )
-        }
-      }
-    }
-
   def showUpdateAddressForm(typ: AddrType): Action[AnyContent] =
     authenticate.async { implicit request =>
       gettingCachedJourneyData[Result](typ) { journeyData =>
