+++ conf/prod.routes
@@ -45,8 +45,8 @@ POST        /personal-account/your-address/:typ/do-you-live-in-the-uk           
 GET         /personal-account/your-address/:typ/cannot-use-the-service                 controllers.AddressController.cannotUseThisService(typ: AddrType)
 GET         /personal-account/your-address/:typ/find-address                           controllers.address.PostcodeLookupController.onPageLoad(typ: AddrType)
 POST        /personal-account/your-address/:typ/find-address                           controllers.address.PostcodeLookupController.onSubmit(typ: AddrType)
-GET         /personal-account/your-address/:typ/select-address                         controllers.AddressController.showAddressSelectorForm(typ: AddrType)
-POST        /personal-account/your-address/:typ/select-address                         controllers.AddressController.processAddressSelectorForm(typ: AddrType)
+GET         /personal-account/your-address/:typ/select-address                         controllers.address.AddressSelectorController.onPageLoad(typ: AddrType)
+POST        /personal-account/your-address/:typ/select-address                         controllers.address.AddressSelectorController.onSubmit(typ: AddrType)
 GET         /personal-account/your-address/:typ/edit-address                           controllers.AddressController.showUpdateAddressForm(typ: AddrType)
 POST        /personal-account/your-address/:typ/edit-address                           controllers.AddressController.processUpdateAddressForm(typ: AddrType)
 GET         /personal-account/your-address/:typ/enter-international-address            controllers.AddressController.showUpdateInternationalAddressForm(typ: AddrType)
