+++ test/controllers/AddressControllerSpec.scala
@@ -166,241 +166,6 @@ class AddressControllerSpec extends BaseSpec with MockitoSugar {
     })
   }
 
-  "Calling AddressController.showAddressSelectorForm" should {
-
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("addressLookupServiceDown" -> Json.toJson(Some(false)), "soleSelectedRecordSet" -> Json.toJson(oneAndTwoOtherPlacePafRecordSet))))
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-
-      val addressLookupResponseFirstPostcode = AddressLookupSuccessResponse(oneAndTwoOtherPlacePafRecordSet)
-      val addressLookupResponseDifferentPostcode = AddressLookupSuccessResponse(newPostcodePlacePafRecordSet)
-    }
-
-    "render the page with the results of the address lookup" when {
-      "RecordSet can be retrieved from the cache" in new LocalSetup {
-        when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-          override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-            block(
-              buildUserRequest(
-                request = FakeRequest("POST", "")
-                  .withFormUrlEncodedBody(
-                    "postcode" -> "AA1 1AA")
-                  .asInstanceOf[Request[A]]
-              )
-                .asInstanceOf[UserRequest[A]]
-            )
-        })
-
-        val result = controller.showAddressSelectorForm(SoleAddrType)(FakeRequest())
-
-        status(result) shouldBe OK
-        verify(mockLocalSessionCache, times(1)).fetch()(any(), any())
-      }
-    }
-
-    "redirect to postcode lookup form" when {
-      "RecordSet can not be retrieved from the cache" in new LocalSetup {
-        when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-          override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-            block(
-              buildUserRequest(
-                request = FakeRequest("POST", "")
-                  .withFormUrlEncodedBody(
-                    "postcode" -> "AA1 1AA")
-                  .asInstanceOf[Request[A]]
-              )
-                .asInstanceOf[UserRequest[A]]
-            )
-        })
-
-        override lazy val sessionCacheResponse: Some[CacheMap] = Some(CacheMap("id", Map.empty))
-
-        val result = controller.showAddressSelectorForm(SoleAddrType)(FakeRequest())
-
-        status(result) shouldBe SEE_OTHER
-        redirectLocation(result) shouldBe Some(address.routes.PostcodeLookupController.onPageLoad(SoleAddrType).url)
-        verify(mockLocalSessionCache, times(1)).fetch()(any(), any())
-      }
-    }
-  }
-
-  "Calling AddressController.processAddressSelectorForm" should {
-
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("addressLookupServiceDown" -> Json.toJson(Some(false)), "postalSelectedRecordSet" -> Json.toJson(oneAndTwoOtherPlacePafRecordSet))))
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-
-      val addressLookupResponseFirstPostcode = AddressLookupSuccessResponse(oneAndTwoOtherPlacePafRecordSet)
-      val addressLookupResponseDifferentPostcode = AddressLookupSuccessResponse(newPostcodePlacePafRecordSet)
-    }
-
-    "call the address lookup service and return 400" when {
-      "supplied no addressId in the form" in new LocalSetup {
-        when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture{
-          override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-            block(
-              buildUserRequest(
-                request = FakeRequest("POST", "")
-                  .withFormUrlEncodedBody(
-                    "postcode" -> "AA1 1AA")
-                  .asInstanceOf[Request[A]]
-              )
-                .asInstanceOf[UserRequest[A]]
-            )
-        })
-
-        when(mockAddressLookupService.lookup(meq("AA1 1AA"), any())(any())) thenReturn {
-          Future.successful(addressLookupResponseFirstPostcode)
-        }
-
-        val result = controller.processAddressSelectorForm(PostalAddrType)(FakeRequest())
-
-        status(result) shouldBe BAD_REQUEST
-        verify(mockLocalSessionCache, times(1)).fetch()(any(), any())
-      }
-
-      "supplied no addressId in the form with a filter" in new LocalSetup {
-
-        when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-          override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-            block(
-              buildUserRequest(
-                request = FakeRequest("POST", "")
-                  .withFormUrlEncodedBody(
-                    "postcode" -> "AA1 1AA",
-                    "filter" -> "7")
-                  .asInstanceOf[Request[A]]
-              )
-                .asInstanceOf[UserRequest[A]]
-            )
-        })
-
-        when(mockAddressLookupService.lookup(any(), any())(any())) thenReturn {
-          Future.successful(addressLookupResponseFirstPostcode)
-        }
-
-        val result = controller.processAddressSelectorForm(PostalAddrType)(FakeRequest())
-
-        status(result) shouldBe BAD_REQUEST
-
-        verify(mockLocalSessionCache, times(1)).fetch()(any(), any())
-      }
-    }
-
-    "call the address lookup service and redirect to the edit address form for a postal address type when supplied with an addressId" in new LocalSetup {
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture{
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = FakeRequest("POST", "")
-                .withFormUrlEncodedBody("addressId" -> "GB990091234514", "postcode" -> "AA1 1AA")
-                .asInstanceOf[Request[A]]
-            )
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      when(mockAddressLookupService.lookup(meq("AA1 1AA"), any())(any())) thenReturn {
-        Future.successful(addressLookupResponseFirstPostcode)
-      }
-
-      val result = controller.processAddressSelectorForm(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/your-address/postal/edit-address")
-      verify(mockLocalSessionCache, times(1))
-        .cache(meq("postalSelectedAddressRecord"), meq(oneOtherPlacePafAddressRecord))(any(), any(), any())
-      verify(mockLocalSessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "call the address lookup service and return a 500 when an invalid addressId is supplied in the form" in new LocalSetup {
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture{
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = FakeRequest("POST", "")
-                .withFormUrlEncodedBody("addressId" -> "GB000000000000", "postcode" -> "AA1 1AA")
-                .asInstanceOf[Request[A]]
-            )
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      when(mockAddressLookupService.lookup(meq("AA1 1AA"), any())(any())) thenReturn {
-        Future.successful(addressLookupResponseFirstPostcode)
-      }
-
-      val result = controller.processAddressSelectorForm(PostalAddrType)(FakeRequest())
-
-      status(result) shouldBe INTERNAL_SERVER_ERROR
-      verify(mockLocalSessionCache, times(0)).cache(any(), any())(any(), any(), any())
-      verify(mockLocalSessionCache, times(1)).fetch()(any(), any())
-    }
-
-    "redirect to enter start date page if postcode is different to currently held postcode" in new LocalSetup {
-      override lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("addressLookupServiceDown" -> Json.toJson(Some(false)), "soleSelectedRecordSet" -> Json.toJson(oneAndTwoOtherPlacePafRecordSet))))
-
-      val cacheAddress = AddressDto.fromAddressRecord(otherPlacePafDifferentPostcodeAddressRecord)
-      val form =
-        FakeRequest("POST", "").withFormUrlEncodedBody("addressId" -> "GB990091234515", "postcode" -> "AA1 2AA")
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = form
-            )
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.processAddressSelectorForm(SoleAddrType)(form)
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/your-address/sole/enter-start-date")
-    }
-
-    "redirect to check and submit page if postcode is not different to currently held postcode" in new LocalSetup {
-      val cacheAddress = AddressDto.fromAddressRecord(twoOtherPlacePafAddressRecord)
-      val form =
-        FakeRequest("POST", "").withFormUrlEncodedBody("addressId" -> "GB990091234515", "postcode" -> "AA1 1AA")
-
-      override lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("addressLookupServiceDown" -> Json.toJson(Some(false)), "soleSelectedRecordSet" -> Json.toJson(oneAndTwoOtherPlacePafRecordSet))))
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture{
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = form
-            )
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-
-      val result = controller.processAddressSelectorForm(SoleAddrType)(form)
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/your-address/sole/changes")
-      verify(controller.sessionCache, times(1))
-        .cache(meq("soleSubmittedStartDateDto"), meq(DateDto(LocalDate.now())))(any(), any(), any())
-    }
-  }
-
   "Calling AddressController.showUpdateAddressForm" should {
 
     trait LocalSetup extends WithAddressControllerSpecSetup {
